Configure cilium with user provided policy enforcement mode
This commit configures cilium to use the user supplied value for policy
enforcement mode, by passing setting it as a value for the helm chart.
The accepted values for this field by cilium are "default", "always" and "never"
If no value is passed, cilium uses "default" as the default value.
When "always" mode is used, cilium blocks all communication.
This can prevent block cluster creation. To prevent this, eks anywhere
will create networkPolicy objects to allow communication between pods crucial
to eks-a. For a management cluster, network policies will be created for all
core kubernetes pods, CAPI, eks-a and flux controllers. For workload cluster,
the network policy will only be created for the kube-system pods.